Check for directory, change default type handling

+++ web_server/server.rb
@@ -1,10 +1,11 @@
 require 'socket'
 require 'json'
 
-content_type = { html: 'text/html',
-                 txt: 'text/plain',
-                 png: 'image/png',
-                 jpg: 'image/jpeg' }
+CONTENT_TYPE = { html:    "text/html",
+                 txt:     "text/plain",
+                 png:     "image/png",
+                 jpg:     "image/jpeg",
+                 default: "application/octet-stream" }
 
 server = TCPServer.open(7680)
 
@@ -14,7 +15,7 @@ loop do
 
   request = socket.gets
   resource = request[/\w+.html/]
-  method = request[/(GET|POST|HEAD)/]
+  method = request[/(GET|POST)/]
   version = request[/HTTP\/\d\.\d/]
   incoming = request[/Content-Length: (\d*)/]
   message = ""
@@ -24,21 +25,23 @@ loop do
   puts "Request: " + request
 
   if method == "GET"
-    if File.exists?("#{resource}")
+    if File.exists?(resource) && !File.directory?(resource)
       ext = File.extname(resource).split(".").last
-      if ext == "html" || ext == "htm" || ext == "txt"
-        file = File.open(resource, "r")
-        file.each do |line|
-          body << line; end
-        status = "200"
-        message = "OK"
-      end
+      if ext == "html" || ext == "htm"
+        type = :html
+      elsif ext == "txt"
+        type = :txt; end
+      file = File.open(resource, "r")
+      file.each do |line|
+        body << line; end
+      status = "200"
+      message = "OK"
     else
       status = "404"
       message = "Not Found"
     end
   elsif method == "POST"
-    if File.exists?("#{resource}")
+    if File.exists?(resource)
       file = File.open(resource, "r")
       file.each do |line|
         #stuff
@@ -56,7 +59,7 @@ loop do
 
   socket.print "#{version} #{status} #{message}\r\n" +
                "Date: " + Time.now.ctime + "\r\n" +
-               "Content-Type: #{content_type[:html]}\r\n" +
+               "Content-Type: #{CONTENT_TYPE.fetch(type, CONTENT_TYPE[:default])}\r\n" +
                "Content-Length: #{body.bytesize}\r\n" +
                "Connection: close\r\n"
   socket.print "\r\n"
